+++ services/madoc-ts/src/frontend/shared/capture-models/new/components/DefaultSubmitButton.tsx
@@ -10,7 +10,7 @@ import { Revisions } from '../../editor/stores/revisions/index';
 import { useDeselectRevision } from '../hooks/use-deselect-revision';
 import { EditorRenderingConfig, useSlotContext } from './EditorSlots';
 
-export const DefaultSubmitButton: EditorRenderingConfig['SubmitButton'] = ({ afterSave }) => {
+export const DefaultSubmitButton: EditorRenderingConfig['SubmitButton'] = ({ afterSave, canSubmit = true }) => {
   const { t } = useTranslation();
   const routeContext = useRouteContext();
   const Slots = useSlotContext();
@@ -93,7 +93,7 @@ export const DefaultSubmitButton: EditorRenderingConfig['SubmitButton'] = ({ aft
                 ) : null}
                 <Button
                   data-cy="publish-button"
-                  disabled={isLoading}
+                  disabled={isLoading || !canSubmit}
                   $primary
                   onClick={() => {
                     saveRevision('submitted').then(() => {